add bourder

+++ Assets/Scripts/Level/GlobalSettings.cs
@@ -8,10 +8,10 @@ namespace Level
         public  static int level = 0;
         public  static  int countObject = 100;
 
-        public  static double x1 = -100;
-        public  static double y1 = 30;
-        public  static double x2 = 100;
-        public  static double y2 = -50;
+        public  static float x1 = -100;
+        public  static float y1 = 30;
+        public  static float x2 = 100;
+        public  static float y2 = -50;
 
         public static List<GameObject> AllObjects = new List<GameObject>();
         public static List<GameObject> AllWear = new List<GameObject>();
